requirements.txt
fitz
Pillow
transformers
scikit-learn
numpy
ultralytics
torch
opencv-python
rapidocr-onnxruntime
sentencepiece